+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- markdownlint-disable MD022 MD024 MD032 -->
 
+## [Unreleased]
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [0.6.0] - 2020-SEP-11
 ### Added
 - Option to clean the output folder can be specified.
@@ -143,6 +152,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Convert Jupyter Notebook to a set of Notebooks, with and without solutions
 - README with instructions for use
 
+[Unreleased]: https://github.com/bryanwweber/thermohw/compare/v0.6.0...HEAD
 [0.6.0]: https://github.com/bryanwweber/thermohw/compare/v0.5.2...v0.6.0
 [0.5.2]: https://github.com/bryanwweber/thermohw/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/bryanwweber/thermohw/compare/v0.5.0...v0.5.1