+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- markdownlint-disable MD022 MD024 MD032 -->
 
-## [Unreleased]
+## [0.7.0] - 2021-JAN-12
 ### Added
 - The `SolutionRemover` preprocessor now uses cell tags to find the solution and its parts
 - A `legacy` command line argument switches back to the old method of finding the solution based on parsing the source of the cell
@@ -160,7 +160,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Convert Jupyter Notebook to a set of Notebooks, with and without solutions
 - README with instructions for use
 
-[Unreleased]: https://github.com/bryanwweber/thermohw/compare/v0.6.0...HEAD
+[0.7.0]: https://github.com/bryanwweber/thermohw/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/bryanwweber/thermohw/compare/v0.5.2...v0.6.0
 [0.5.2]: https://github.com/bryanwweber/thermohw/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/bryanwweber/thermohw/compare/v0.5.0...v0.5.1